requirements.txt
numpy==1.26.4
tensorflow==2.16.0rc0
Pillow==10.2.0
torch==2.2.1
torchvision==0.17
scipy==1.12.3
scikit-learn==1.4.1
opencv-python==4.9.0.80